Adds QueryLogTrait
Removes silly return statement
